backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
sentence-transformers
scikit-learn
langgraph
langchain-core
icalendar
groq
